### 2025-05-22 #8257 @editorialbot
[PRE REVIEW]: DiRe - JAX: A JAX based Dimensionality Reduction Algorithm for Large-scale Data
**Submitting author:** @sashakolpakov (<a href="http://orcid.org/0000-0002-6764-8894">Alexander Kolpakov</a>)
**Repository:** https://github.com/sashakolpakov/dire-jax
**Branch with paper.md** (empty if default branch): dire-jax-joss
**Version:** 0.0.1-alpha
