### 2025-05-22 #8257 @sashakolpakov
@danielskatz Thanks for the comments!

"1. You say your software version is 0.0.1-alpha ... " Yes, I do, mainly because I never thought it could be a decisive factor. The software has now >100 commits, > 3k lines of code (both .py package files + .ipynb notebooks) combined, and it has been accepted to the institutional PyPI index (https://pypi.org/org/UATX/). I don't think that it is immature, as it is an installable and fully functional package that has both GPU and TPU (therefore grants from Google) acceleration available. It also needs future work, but a two-men army cannot plausibly do everything. If there is some interest from the community, it would definitely help. On my side, I would suggest @lmcinnes to be the right person to ask, as we are truing to address some (apparently, fundamental) limitation of UMAP that were recently pointed out by Pachter et al. (https://doi.org/10.1371/journal.pcbi.1011288). Also, the JAX framework developed in DiRe-JAX may be used to port other dimension reducers to JAX, such as UMAP, tSNE. Even though they've been ported by now by rapids.ai, certain classical datasets (MNIST Digits) did not have their global structure preserved by the GPU implementation (at the moment of writing, https://github.com/rapidsai/cuml/issues/5474; https://github.com/rapidsai/cuml/issues/5782). However, our main goal is to make a cross-platform dimension reducer that addresses some fundamental limitations of SOTA algorithms, not just fix some bugs.

"2. our paper looks very long ..." It does. I think the best way to make it shorter is to remove all the technical details with how we intend to preserve the global structure of the dataset, and how we measure it. I can make a reference to the arXiv pre-print that has all the details [arXiv:2503.03156](https://arxiv.org/abs/2503.03156). I'm gonna do the edits and then try to let the editorial bot get another stab at it.

Hope for a productive discussion, and hope there will be some interest in JAXifying good things!
