+++ paper/paper.md
@@ -39,7 +39,7 @@ This makes it suitable for a wide range of applications in machine learning, bio
 Traditional dimensionality reduction techniques such as UMAP and tSNE are widely used for visualizing high-dimensional data in lower-dimensional spaces, 
 usually 2D and sometimes 3D. Other uses include dimensionality reduction to other, possibly higher and thus non-visual dimensions, for the subsequent use of classifiers such as SVMs.
 
-However, these methods often struggle with scalability, interpretability, and preservation of global data structures. UMAP, while fast and scalable, may overemphasize 
+However, these methods often struggle with scalability, interpretability, and preservation of global data structures. For example, while fast and scalable, UMAP may overemphasize 
 local structures at the expense of global data relationships [@pachter]. And tSNE, while known for producing high-quality visualizations, may be computationally 
 expensive and sensitive to hyperparameter tuning [@kobak2019art].
 
@@ -51,6 +51,35 @@ DiRe-JAX also includes a wealth of metrics for analyzing embedding quality and f
 hyperparameter optimization become feasible even in low-cost environments like Google Colab.
 
 This makes DiRe-JAX an essential toolkit for researchers and practitioners working with complex, high-dimensional data.
+
+# Benchmarks
+
+A few benchmark below provide some visuals regarding the global structure changes or preservation by various methods, such as 
+
+## Dataset: Disk Uniform
+
++-------------------------------------------------------------------------------------+------------------------------------------------------------------------------------+
+| ![Disk DiRe–JAX embedding](../pics/embeddings/disk-dire-jax.png){height="80pt"}     | ![Disk tSNE embedding](../pics/embeddings/disk-tsne.png){height="80pt"}            |
++-------------------------------------------------------------------------------------+------------------------------------------------------------------------------------+
+| ![Disk cuML UMAP embedding](../pics/embeddings/disk-cuml-umap.png){height="80pt"}   | ![Disk UMAP embedding](../pics/embeddings/disk-umap.png){height="80pt"}            |
++-------------------------------------------------------------------------------------+------------------------------------------------------------------------------------+
+
+## Dataset: Two Half–Moons
+
++-------------------------------------------------------------------------------------+------------------------------------------------------------------------------------+
+| ![Moons DiRe–JAX embedding](../pics/embeddings/moons-dire-jax.png){height="80pt"}   | ![Moons tSNE embedding](../pics/embeddings/moons-tsne.png){height="80pt"}          |
++-------------------------------------------------------------------------------------+------------------------------------------------------------------------------------+
+| ![Moons cuML UMAP embedding](../pics/embeddings/moons-cuml-umap.png){height="80pt"} | ![Moons UMAP embedding](../pics/embeddings/moons-umap.png){height="80pt"}          |
++-------------------------------------------------------------------------------------+------------------------------------------------------------------------------------+
+
+
+## Dataset: Levine 32
+
++--------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------+
+| ![Levine 32 DiRe–JAX embedding](../pics/embeddings/levine32-dire-jax.png){height="80pt"}   | ![Levine 32 tSNE embedding](../pics/embeddings/levine32-tsne.png){height="80pt"}   |
[…]
