+++ paper/paper.bib
@@ -165,6 +165,7 @@
   pages={561--580},
   year={2007},
   publisher={IOS Press},
+  doi={10.3233/ida-2007-11508},
   url={https://doi.org/10.3233/IDA-2007-11508}
 }
 
