### 2025-05-22 #8257 @danielskatz
Our bot only reads the first command in a comment...
