### 2025-05-22 #8257 @danielskatz
Thanks.  Let's go ahead with this for now.  I'll look for an editor next.

Also, can you change the title entry in your `JL` references to `  title={Extensions of {L}ipschitz mappings into a {H}ilbert space},`?
