+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{kobak2019art,
-  title={The art of using t-SNE for single-cell transcriptomics},
+  title={The art of using t-{SNE} for single-cell transcriptomics},
   author={Kobak, Dmitry and Berens, Philipp},
   journal={Nature Communications},
   volume={10},
@@ -14,7 +14,7 @@
 @article{pachter,
   title={The specious art of single-cell genomics},
   author={Chari, Tara and Pachter, Lior},
-  journal={PLoS Computational Biology},
+  journal={{PL}o{S} {C}omputational {B}iology},
   volume={19},
   number={8},
   pages={e1011288},
@@ -57,7 +57,7 @@
 
 @misc{umap-git,
   author       = {{UMAP} Developers},
-  title        = {Uniform Manifold Approximation and Projection {(UMAP)}},
+  title        = {{U}niform manifold approximation and projection {(UMAP)}},
   year         = {2025},
   publisher    = {GitHub},
   journal      = {GitHub Repository},
@@ -66,7 +66,7 @@
 
 @misc{persim-docs,
   author       = {{Persim Developers}},
-  title        = {Persim: Distances and representations of persistence diagrams},
+  title        = {{P}ersim: distances and representations of persistence diagrams},
   year         = {2024},
   publisher    = {GitHub},
   journal      = {GitHub Repository},
@@ -75,7 +75,7 @@
 
 @misc{pytwed-docs,
   author       = {Marcus Vo\ss and Jonathan Zumer},
-  title        = {PyTWED: Python wrapper for Time Warped Edit Distance},
+  title        = {{P}y{TWED}: {P}ython wrapper for time warped edit distance},
   year         = {2021},
   publisher    = {GitHub},
   journal      = {GitHub Repository},
@@ -97,7 +97,7 @@
 
 @misc{dire-github,
   author       = {Alexander Kolpakov and Igor Rivin},
-  title        = {DiRe: Dimensionality Reducer},
+  title        = {{D}i{R}e: dimensionality reducer},
   year         = {2025},
   publisher    = {GitHub},
   journal      = {GitHub Repository},
@@ -105,7 +105,7 @@
 }
 
 @misc{dire-arxiv,
-      title={{DiRe-JAX: A JAX based Dimensionality Reduction Algorithm for Large-scale Data}},
+      title={{D}i{R}e-{JAX}: A {JAX} based dimensionality reduction algorithm for large-scale data},
       author = {Alexander Kolpakov and Igor Rivin},
       year={2025},
       eprint={arXiv:2503.03156},
@@ -114,7 +114,7 @@
 }
 
 @article{flamary-pot,
-  title={POT: Python Optimal Transport},
+  title={{POT}: {P}ython optimal transport},
   author={Flamary, R{\'e}mi and Courty, Nicolas and Gramfort, Alexandre and Alaya, Mokhtar Z. and Boisbunon, Aur{\'e}lie and Chambon, Stanislas and Chapel, Laetitia and Corenflos, Adrien and Fatras, Kilian and Fournier, Nemo and Gautheron, Loris and Gayraud, Nathalie T.H. and Janati, Hicham and Rakotomamonjy, Alain and Redko, Ievgen and Rolet, Antoine and Schutz, Antony and Seguy, Vivien and Sutherland, Danica J. and Tavenard, Romain and Tong, Alexander and Vayer, Titouan},
   journal={Journal of Machine Learning Research},
   volume={22},
@@ -126,7 +126,7 @@
 
 @misc{ot-docs,
   author       = {{POT Developers}},
-  title        = {POT: Python Optimal Transport},
+  title        = {{POT}: {P}ython optimal transport},
   year         = {2025},
   publisher    = {GitHub},
[…]
