### 2025-05-22 #8257 @sashakolpakov
> Also, can you change the title entry in your `JL` references to ` title={Extensions of {L}ipschitz mappings into a {H}ilbert space},`?

@danielskatz Done.
