### 2025-05-22 #8257 @crhea93
Yes I can too

Merci Beaucoup,
  Carter Rhea
  Doctorat en astrophysique
  Développeur de logiciels chez Dragonfly
  https://crhea93.github.io/
