### 2025-05-28 #8264 @crhea93
For the documentation, I see that there are documents in the `docs` folder, but they do not appear to be hosted anywhere. Could you make that available? I'd also suggest adding a GitHub action to automatically build the docs after each PR or push to master.
