### 2025-05-30 sashakolpakov/dire-jax#16 @crhea93
Contributing section
Thanks for the very clear README file! Would it be possible to include more content in the Contributing section?

Here is an example template:
Contributing to [Project Name]

Thank you for considering contributing to [Project Name]! We welcome contributions from everyone. By participating in this project, you agree to abide by our [Code of Conduct](https://github.com/openjournals/joss-reviews/issues/CODE_OF_CONDUCT.md).
How Can I Contribute?
Reporting Bugs

    Ensure the bug was not already reported by searching on GitHub under [Issues](https://github.com/yourusername/yourproject/issues).
    If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/yourusername/yourproject/issues/new). Be sure to include a title and clear description, as much relevant information as possible, and a code sample or an executable test case demonstrating the expected behavior that is not occurring.

Suggesting Enhancements

    Open a new issue with a clear title and detailed description.
    Provide a step-by-step description of the suggested enhancement in as many details as possible.
    Provide specific examples to demonstrate the steps.

Your First Code Contribution

    Fork the repository and create your branch from main.
    If you've added code that should be tested, add tests.
    Ensure the test suite passes.
    Make sure your code lints.

Pull Requests

    Fill in the required template.
    Do not include issue numbers in the PR title.
    Include screenshots and animated GIFs in your pull request whenever possible.
    Follow the [style guide](https://github.com/openjournals/joss-reviews/issues/8264#style-guides).
    Document new code based on the [Documentation Styleguide](https://github.com/openjournals/joss-reviews/issues/8264#documentation-styleguide).
    End all files with a newline.

Style Guides
Git Commit Messages

    Use the present tense ("Add feature" not "Added feature").
    Use the imperative mood ("Move cursor to..." not "Moves cursor to...").
    Limit the first line to 72 characters or less.
    Reference issues and pull requests liberally after the first line.

[Language] Styleguide

    Follow the existing coding style present in the codebase.
    [Additional specific style guidelines if any]

Documentation Styleguide

    Use [Markdown](https://daringfireball.net/projects/markdown).
    Reference methods and classes in markdown with the custom {} notation:
        Reference classes with {ClassName}
        Reference instance methods with {ClassName::methodName}
        Reference class methods with {ClassName.methodName}

Additional Notes
Issue and Pull Request Labels

This section lists the labels we use to help us track and manage issues and pull requests.

[Describe the labels and their purposes]

Thank you for contributing to [Project Name]!

Speaking of templates, would it be possible to make a template for new GitHub issues or PRs? They have been very helpful in the past :)
