### 2025-05-30 sashakolpakov/dire-jax#16 @sashakolpakov
Addressed in README.md with a detailed contributing guide in the documentation.
