### 2025-05-30 sashakolpakov/dire-jax#16 @sashakolpakov
@crhea93 Thanks! Issue is addressed in README.md with a detailed contributing guide in the documentation.
