### 2025-06-06 #8264 @Treys925
@crhea93 thanks for raising those issues, I don't see anything else that requires fixing prior to publication.
@sashakolpakov this is great work, we need better alternatives to UMAP and t-sne.
@nkrusch I'm happy to recommend this paper for publication in JOSS.
