### 2025-06-06 #8264 @nkrusch
**Ref: Comments on  the [Post-Review Checklist](https://github.com/openjournals/joss-reviews/issues/8264#issuecomment-2950624611)**

Some comments that are worth mentioning, but not issues.

#### About the text of the paper

During pre-review, there was an issue around paper length, which was then addressed. The paper is still a bit long. For example, the sections on embeddings could be removed, or moved to the documentation, without conceptual loss. However, I am not raising an issue since the length is overall (borderline) acceptable in comparison to similar submissions.

I generally recommend adding to the documentation an "Introduction" section, that explains what kind of problems the DiRe - JAX can address, and how it works conceptually. The discussions on embeddings could also go there.

#### Check title, author list, version tag, and the license

The software version is currently 0.1.0, which is OK. However, the [semantics of that version number](https://semver.org/#semantic-versioning-specification-semver) is:

> Major version zero (0.y.z) is for initial development. Anything MAY change at any time. The public API SHOULD NOT be considered stable.

If this meaning accurately reflects the development stage, then it should stay at that version.
