### 2025-06-06 #8264 @sashakolpakov
@nkrusch Thank you for the comments!

1. If I may, I would keep the changes to the paper minimal. It is already a very abridged version of the arXiv manuscript, and I would like to make a version of record for JOSS that contains most of the essential details rather than turn it into a smaller abstract-like version that refers to an arXiv preprint (and the latter can be easily changed, appended, etc).

2. The version here (0.1.0) is mature enough to satisfy the reviewers. However, following a comment of @crhea93, @lmcinnes was kind enough to provide more instructions on the desirable features. I would like to follow up, though it will require work, before I can call it a genuinely stable version (just to be on the safe side and not make a wrong impression on the potential users).

3. The four bullet points about archiving the accepted version will be completed in the following days.

Finally, I would like to thank everyone: @danielskatz, @nkrusch, @crhea93, @Treys925 and @lmcinnes for their great help!
