### 2025-06-07 #8264 @sashakolpakov
Please find the requested info below.

1. Release tag [joss-0.2.0](https://github.com/sashakolpakov/dire-jax/releases/tag/joss-0.2.0)
2. [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15611800.svg)](https://doi.org/10.5281/zenodo.15611800)
3. The paper had some typos, now corrected. Added thanks to the reviewers and discussion participants.

Thank you, @nkrusch, and the JOSS editorial team!
