### 2025-06-07 #8264 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
