### 2025-06-09 bdef80a
Update paper.md
