### 2025-06-09 #8264 @danielskatz
👋 @sashakolpakov - As the track editor, I've proofread and checked this submission, and request that you make 2 minor changes in the paper, as shown in https://github.com/sashakolpakov/dire-jax/pull/23. Otherwise, this all looks good and is ready to accept.
