### 2025-06-09 #8264 @sashakolpakov
@danielskatz I did some cosmetic edits to two (2) equations / formulas. Please check the diff (https://github.com/sashakolpakov/dire-jax/pull/24). If this is acceptable, then I will merge, and it is ready to go. Thanks!
